% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_numeric_function.R
\name{chr_numeric}
\alias{chr_numeric}
\title{Chromosome Numeric}
\usage{
chr_numeric(Chr)
}
\arguments{
\item{Chr}{a character column or vector of chromosome names}
}
\description{
Convert chromosome names from character to numeric.
Useful when sorting a dataframe by chromosome number.
}
\keyword{chromosome}
\keyword{numeric}
