% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_sort_function.R
\name{chr_sort}
\alias{chr_sort}
\title{Chromosome Sort}
\usage{
chr_sort(Chr)
}
\arguments{
\item{Chr}{character vector of chromosome names}
}
\value{
A character vector, sorted in ascending order of chromosome number
}
\description{
Sort chromosomes in ascending order
}
\examples{
Chr=c("chr3","chr7","chr6","chr2")
chr_sort(Chr)
}
\keyword{chromosome}
\keyword{sort}
