% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_function.R
\name{clean_data}
\alias{clean_data}
\title{Clean Data}
\usage{
clean_data(snp.info, peak.info)
}
\arguments{
\item{snp.info}{SNP information dataframe}

\item{peak.info}{Peak annotation dataframe}
}
\description{
Clean snp.info and peak.info data and save as RData file
}
\examples{
clean_data_function()
}
\keyword{SNP}
\keyword{clean}
\keyword{data}
\keyword{peak}
