% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data_spc_function.R
\name{combine_data_spc}
\alias{combine_data_spc}
\title{Combine Data SNP-Peak-Clinical}
\usage{
combine_data_spc(chr)
}
\arguments{
\item{chr}{chromosome name}
}
\description{
SNP-Peak-Clinical data: combine projid, SNPs, PCs, membership,
  age, gender, peaks, and outcome data for a chromosome
}
\keyword{dataprep}
\keyword{spc}
