% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data_spg_function.R
\name{combine_data_spg}
\alias{combine_data_spg}
\title{Combine Data SNP-Peak-Gene}
\usage{
combine_data_spg(chr)
}
\arguments{
\item{chr}{chromosome name}
}
\description{
SNP-Peak-Gene Data: combine projid, SNPs, PCs, membership,
  age, gender, peaks, gene and outcome data for a chromosome.
}
\keyword{combine}
\keyword{data}
\keyword{gene}
