% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_by_chr_function.R
\name{data_by_chr}
\alias{data_by_chr}
\title{Data by Chromosome}
\usage{
data_by_chr(chrs, snps, peaks)
}
\arguments{
\item{chrs}{Character vector of chromosomes names}

\item{snps}{SNP information dataframe}

\item{peaks}{Peak information dataframe}
}
\description{
save snp.info, peak.info for each chromosome
   as "data/chipseq_\code{chr}.RData"
}
\keyword{data}
