% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_spc_function.R
\name{data_prep_spc}
\alias{data_prep_spc}
\title{SNP-Peak-Clinical Data Prep Helper}
\usage{
data_prep_spc(chr)
}
\arguments{
\item{chr}{chromosome name}
}
\description{
Helper function for SNP-Peak-Clinical mediation data. 
Combine projid, SNPs, PCs, membership, age, gender, peaks, and outcome data for a chromosome
}
\keyword{data}
