% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_data_function.R
\name{gene_data}
\alias{gene_data}
\title{Gene Data Prep}
\usage{
gene_data(chr)
}
\arguments{
\item{chr}{chromosome name}
}
\description{
Gene data prep for a chromosome
}
\examples{
gene_data_function()
}
\keyword{data}
\keyword{gene}
\keyword{prep}
