% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mediation_table_function.R
\name{gene_mediation_table}
\alias{gene_mediation_table}
\title{Gene Mediation Table}
\usage{
gene_mediation_table(med_dat, match_row, simulations)
}
\arguments{
\item{med_dat}{gene mediation data}

\item{match_row}{row in match dataframe: col1 SNP, col2 peak, col3 gene}

\item{simulations}{number of simulations to run}
}
\description{
Create a mediation table for one match in the gene data
}
\examples{
gene_mediation_table_function()
}
\keyword{mediation}
\keyword{table}
