% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mediation_table_chr_function.R
\name{gene_mediation_table_chr}
\alias{gene_mediation_table_chr}
\title{Gene Mediation Table Chromosome}
\usage{
gene_mediation_table_chr(chr, matches_by_gene, simulations)
}
\arguments{
\item{chr}{chromosome name}

\item{matches_by_gene}{list of peak-SNP for each gene}

\item{simulations}{number of simulations to run}
}
\description{
Create a mediation table for an entire chromosome of gene data
}
\examples{
gene_mediation_table_chr_function()
}
\keyword{gene}
\keyword{mediation}
\keyword{table}
