% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_res_function.R
\name{gene_res}
\alias{gene_res}
\title{Gene Results Prep}
\usage{
gene_res(chr, med.data, matches)
}
\arguments{
\item{chr}{chromosome name}

\item{med.data}{gene mediation data}

\item{matches}{gene matches dataframe}
}
\description{
Mediation analysis gene data prep for a chromosome
}
\examples{
gene_res_function()
}
\keyword{data}
\keyword{gene}
\keyword{mediation}
\keyword{prep}
