% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneticMediation.R
\docType{package}
\name{GeneticMediation-package}
\alias{GeneticMediation}
\alias{GeneticMediation-package}
\title{Genetic Mediation: An R Package for Causal Mediation Analysis of ROSMAP Data}
\description{
GeneticMediation was created as part of the UNC Wu Lab project
Constructing Causal Mediation Pathways For Multi-Level Genomic Data.
GeneticMediation provides methods for cleaning and processing data from
the ROSMAP study and preparing it for causal mediation analysis
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tydarnell/GeneticMediation}
  \item \url{http://www.tydarnell.com/GeneticMediation}
}

}
\author{
\strong{Maintainer}: Ty Darnell \email{tydarnell@gmail.com}

}
\keyword{internal}
