% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneticMediation.R
\docType{package}
\name{GeneticMediation}
\alias{GeneticMediation}
\title{GeneticMediation: A package for conducting causal mediation analysis on ROSMAP data}
\description{
GeneticMediation provides methods for conducting causal mediation analysis on data from the ROSMAP study.
It also provides methods for cleaning, matching, and preparing the data for analysis.
}
