% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_to_first_function.R
\name{last_to_first}
\alias{last_to_first}
\title{Last to First}
\usage{
last_to_first(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
Make the last column the first column
}
\examples{
last_to_first_function()
}
\keyword{column}
\keyword{first}
\keyword{last}
