% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu_function.R
\name{lu}
\alias{lu}
\title{Length Unique}
\usage{
lu(x)
}
\arguments{
\item{x}{vector}
}
\description{
Get the length of unique values of a vector
}
\examples{
lu_function()
}
\keyword{length}
\keyword{unique}
