% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_folder_function.R
\name{make_folder}
\alias{make_folder}
\title{Make Folder}
\usage{
make_folder(path)
}
\arguments{
\item{path}{path for new folder}
}
\description{
Make a folder only if the folder does not exist
}
\examples{
make_folder_function()
}
\keyword{folder}
