% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_all_function.R
\name{match_all}
\alias{match_all}
\title{Match SNPs Peaks All Chromosomes}
\usage{
match_all(snp.info, peak.info, chrs)
}
\arguments{
\item{snp.info}{SNP information}

\item{peak.info}{Peak information}

\item{chrs}{character vector of chromosome names}
}
\description{
Match SNPs and Peaks in all chromosomes and 
  return dataframe of matches
}
\keyword{SNP}
\keyword{match}
\keyword{peak}
