% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_allchrs_function.R
\name{match_allchrs}
\alias{match_allchrs}
\title{Match all Chromosomes}
\usage{
match_allchrs(snp.info, peak.info, chrs)
}
\arguments{
\item{snp.info}{SNP information}

\item{peak.info}{Peak information}

\item{chrs}{character vector of chromosome names}
}
\description{
Match Peaks and SNPs in a character vector of chromosome names
}
\examples{
match_allchrs_function()
}
\keyword{SNP}
\keyword{match}
\keyword{peak}
