% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_chr_function.R
\name{match_chr}
\alias{match_chr}
\title{Match Chromosome}
\usage{
match_chr(snp.info, peak.info)
}
\arguments{
\item{snp.info}{SNP information}

\item{peak.info}{Peak information}
}
\description{
Match Peaks and SNPs in a chromosome
}
\keyword{SNP}
\keyword{match}
\keyword{peak}
