% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_snp_peak_function.R
\name{match_snp_peak}
\alias{match_snp_peak}
\title{Match SNPs and Peaks}
\usage{
match_snp_peak(snp.info, peak.info)
}
\arguments{
\item{snp.info}{SNP information for chr}

\item{peak.info}{Peak information chr}
}
\description{
Match SNPs and Peaks in a chromosome
}
\keyword{match}
