% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med.res_spg_function.R
\name{med.res_spg}
\alias{med.res_spg}
\title{SNP-Peak-Gene Mediation Data Prep}
\usage{
med.res_spg(chr, med.data, matches, mediator.path)
}
\arguments{
\item{chr}{chromosome name}

\item{med.data}{gene mediation data}

\item{matches}{SNP-Peak-Gene matches dataframe}

\item{mediator.path}{path to mediator folder with mediation_chr#_new.RData files}
}
\description{
Prepare a chromosome SNP-Peak-Gene data for mediation analysis data
and save as "data/gene.res/res_\code{chr}.RData"
}
\keyword{data}
