% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_all_spc_function.R
\name{med_all_spc}
\alias{med_all_spc}
\title{SNP-Peak-Clinical Mediation Table}
\usage{
med_all_spc(chrs, covar, simulations)
}
\arguments{
\item{chrs}{character vector of chromosome names}

\item{covar}{covariates string, each covariate separated by +}

\item{simulations}{number of simulations to run}
}
\description{
Create a table of SNP-Peak-Clinical mediation results
 for all chromosomes
}
\keyword{mediation}
\keyword{table}
