% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_chr_spc_function.R
\name{med_chr_spc}
\alias{med_chr_spc}
\title{SNP-Peak-Clinical Mediation Table Chromosome}
\usage{
med_chr_spc(matches, med.res.data, covar, simulations)
}
\arguments{
\item{matches}{SNP-Peak matches dataframe for a chromosome}

\item{med.res.data}{mediation data}

\item{covar}{covariates string, separate covariates with +}

\item{simulations}{number of simulations to run}
}
\description{
Create SNP-Peak-Clinical mediation table for SNP-Peak matches in a chromosome
}
\keyword{mediation}
\keyword{table}
