% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_chr_spg_function.R
\name{med_chr_spg}
\alias{med_chr_spg}
\title{SNP-Peak-Gene Mediation Table Chromosome}
\usage{
med_chr_spg(chr, gene_matches, covar, simulations)
}
\arguments{
\item{chr}{chromosome name}

\item{gene_matches}{list of peak-SNP matches for each gene}

\item{covar}{covariates string, separate covariates with +}

\item{simulations}{number of simulations to run}
}
\description{
Create a SNP-Peak-Gene mediation table for a chromosome
}
\keyword{mediation}
\keyword{table}
