% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_data_prep_function.R
\name{med_data_prep}
\alias{med_data_prep}
\title{Mediation Data Prep}
\usage{
med_data_prep(chr)
}
\arguments{
\item{chr}{chromosome name}
}
\description{
Combine projid, SNPs, PCs, membership, age, gender, peaks, and outcome data for a chromosome
}
\examples{
med_data_prep_function()
}
\keyword{data}
\keyword{mediation}
\keyword{prep}
