% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_res_data_function.R
\name{med_res_data}
\alias{med_res_data}
\title{Mediation Result Data Prep}
\usage{
med_res_data(chr)
}
\arguments{
\item{chr}{chromosome name}
}
\description{
Mediation analysis data prep for a chromosome
}
\examples{
med_res_data_function()
}
\keyword{data}
\keyword{mediation}
\keyword{prep}
