% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_res_data_save_function.R
\name{med_res_data_save}
\alias{med_res_data_save}
\title{Mediation Result Data Save}
\usage{
med_res_data_save(chrs, all_matches)
}
\arguments{
\item{chrs}{character vector of chromosome names}

\item{all_matches}{dataframe of peak-SNP matches for all chromosomes}
}
\description{
Save Mediation analysis data and matches for each chromosome
}
\examples{
med_res_data_save_function()
}
\keyword{data}
\keyword{mediation}
\keyword{result}
\keyword{save}
