% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_results_spg_function.R
\name{med_results_spg}
\alias{med_results_spg}
\title{Mediation Results SNP-Peak-Gene}
\usage{
med_results_spg(med.res.data, match_row)
}
\arguments{
\item{med.res.data}{mediation data}

\item{match_row}{row of SNP-Peak-Gene matches}
}
\description{
Obtain Mediation Results for Sensitivity Analysis
}
