% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_summary_spg_function.R
\name{med_summary_spg}
\alias{med_summary_spg}
\title{Mediation Summary SNP-Peak-Gene}
\usage{
med_summary_spg(med.res.data, match_row, simulations = 1000)
}
\arguments{
\item{med.res.data}{gene mediation data}

\item{match_row}{row in match dataframe: col1 SNP, col2 Peak, col3 Gene}

\item{simulations}{number of simulations, 1000 by default}
}
\description{
Mediation summary for one SNP-Peak-Gene match
}
\keyword{mediation}
\keyword{summary}
