% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_table_spc_function.R
\name{med_table_spc}
\alias{med_table_spc}
\title{SNP-Peak-Clinical Mediation Table}
\usage{
med_table_spc(med.res.data, match_row, covar, simulations)
}
\arguments{
\item{med.res.data}{mediation data}

\item{match_row}{row in match dataframe, should have 2 columns}

\item{covar}{covariates character variable, separate covariates with +}

\item{simulations}{number of simulations to run}
}
\description{
Create SNP-Peak-Clinical mediation table for a SNP-Peak match
}
\keyword{mediation}
\keyword{table}
