% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_spg_function.R
\name{mediation_spg}
\alias{mediation_spg}
\title{Mediation SNP-Peak-Gene}
\usage{
mediation_spg(med.res.data, match_row, simulations = 1000)
}
\arguments{
\item{med.res.data}{gene mediation data}

\item{match_row}{row in match dataframe: col1 SNP, col2 Peak, col3 Gene}

\item{simulations}{number of simulations, 1000 by default}
}
\description{
Mediation analysis one SNP-Peak-Gene match
}
\keyword{mediation}
\keyword{summary}
