% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_table_function.R
\name{mediation_table}
\alias{mediation_table}
\title{Mediation Table}
\usage{
mediation_table(med.res.data, match_row, covar, outcome, simulations)
}
\arguments{
\item{med.res.data}{mediation data}

\item{match_row}{row in match dataframe, should have 2 columns}

\item{covar}{covariates character variable, separate covariates with +}

\item{outcome}{outcome character variable}

\item{simulations}{number of simulations to run}
}
\description{
Create a mediation table for one peak-SNP match
}
\examples{
mediation_table_function()
}
\keyword{mediation}
\keyword{table}
