% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_table_allchr_function.R
\name{mediation_table_allchr}
\alias{mediation_table_allchr}
\title{Mediation Table all Chromosomes}
\usage{
mediation_table_allchr(chrs, folderpath, covar, outcome, simulations)
}
\arguments{
\item{chrs}{character vector of chromosome names}

\item{folderpath}{path to mediation results folder}

\item{covar}{covariates character variable, separate covariates with +}

\item{outcome}{outcome variable string}

\item{simulations}{number of simulations to run}
}
\description{
Create a table of mediation results for a character vector of chromosome names
}
\examples{
mediation_table_allchr_function()
}
\keyword{mediation}
\keyword{table}
