% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_table_chr_function.R
\name{mediation_table_chr}
\alias{mediation_table_chr}
\title{Mediation Table Chromosome}
\usage{
mediation_table_chr(matches, med.res.data, covar, outcome, simulations)
}
\arguments{
\item{matches}{SNP-peak matches dataframe for a chromosome}

\item{med.res.data}{mediation data}

\item{covar}{covariates string, separate covariates with +}

\item{outcome}{outcome character variable}

\item{simulations}{number of simulations to run}
}
\description{
Create a table of mediation results for all peak-SNP matches in a chromosome
}
\examples{
mediation_table_chr_function()
}
\keyword{mediation}
\keyword{table}
