% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_csv_function.R
\name{merge_csv}
\alias{merge_csv}
\title{Merge CSV}
\usage{
merge_csv(path)
}
\arguments{
\item{path}{path to folder containing csv files}
}
\description{
merge multiple csv files into a single dataframe
}
\keyword{helper}
