% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_summary_function.R
\name{sens_summary}
\alias{sens_summary}
\title{Sensitivity Summary}
\usage{
sens_summary(res, i)
}
\arguments{
\item{res}{mediation results dataframe}

\item{i}{row index of dataframe}
}
\description{
Sensitivity Analysis Summary
}
