% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_chr_spg_function.R
\name{sig_chr_spg}
\alias{sig_chr_spg}
\title{Significant SNP-Peak-Gene}
\usage{
sig_chr_spg(chr)
}
\arguments{
\item{chr}{chromosome name}
}
\description{
Returns summary of mediation results for a chr with significant results
}
\keyword{mediation}
\keyword{spg}
