% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_spg_function.R
\name{sig_spg}
\alias{sig_spg}
\title{Significant SNP-Peak-Gene}
\usage{
sig_spg(sig_dat, row)
}
\arguments{
\item{sig_dat}{significant results dataframe}

\item{position}{row in data frame}
}
\description{
Returns summary of mediation results for a row in significant data frame
}
\keyword{mediation}
\keyword{spg}
