% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_peak_bychr_function.R
\name{snp_peak_bychr}
\alias{snp_peak_bychr}
\title{SNPs and Peaks by Chromosome}
\usage{
snp_peak_bychr(chrs, snps, peaks)
}
\arguments{
\item{chrs}{Character vector of chromosomes names}

\item{snps}{SNP information dataframe}

\item{peaks}{Peak information dataframe}
}
\description{
save snp.info, peak.info for each chromosome
   as "data/chipseq/chipseq_\code{chr}.RData"
}
\keyword{data}
