% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_readcount_function.R
\name{transpose_readcount}
\alias{transpose_readcount}
\title{Transpose Readcount}
\usage{
transpose_readcount(readcount)
}
\arguments{
\item{readcount}{ChIP-seq readcount dataframe}
}
\description{
Transpose ChIP-seq Reacount dataframe and match project id
}
\examples{
transpose_readcount_function()
}
\keyword{readcount}
\keyword{transpose}
