% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_function.R
\name{clean_data}
\alias{clean_data}
\title{Clean Data}
\usage{
clean_data(snp.path, peak.path)
}
\arguments{
\item{snp.path}{path to SNP information dataframe}

\item{peak.path}{path to Peak annotation dataframe}
}
\description{
Clean snp.info, peak.info data and save as data/chipseq.RData
}
\keyword{dataprep}
\keyword{spc}
