% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_spg_function.R
\name{data_prep_spg}
\alias{data_prep_spg}
\title{SNP-Peak-Gene Data Prep Helper}
\usage{
data_prep_spg(chr, mediator.path)
}
\arguments{
\item{chr}{chromosome name}

\item{mediator.path}{path to mediator folder with mediation_chr#_new.RData files}
}
\description{
Helper function for SNP-Peak-Gene data prep.
   Combine projid, SNPs, PCs, membership, age, gender, peaks, gene and 
   outcome data for a chromosome.
}
\keyword{data}
\keyword{prep}
