% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{GeneticMediation}
\alias{GeneticMediation}
\alias{GeneticMediation-package}
\title{GeneticMediation: A package for conducting causal mediation analysis on ROSMAP data}
\description{
GeneticMediation provides methods for conducting causal mediation analysis on data from the ROSMAP study.
It also provides methods for cleaning, matching, and preparing the data for analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tydarnell.github.io/GeneticMediation}
  \item \url{https://github.com/tydarnell/GeneticMediation}
}

}
\author{
\strong{Maintainer}: Ty Darnell \email{tydarnell@gmail.com}

}
