% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med.res_spc_function.R
\name{med.res_spc}
\alias{med.res_spc}
\title{SNP-Peak-Clinical Mediation Data Prep}
\usage{
med.res_spc(chr, all_matches)
}
\arguments{
\item{chr}{chromosome name}

\item{all_matches}{dataframe of SNP-Peak matches for all chromosomes}
}
\description{
Prepare a chromosome of SNP-Peak-Clinical data for mediation analysis
and save as "data/spc.res/res_\code{chr}.RData"
}
\keyword{data}
\keyword{prep}
