% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_spg_function.R
\name{sig_spg}
\alias{sig_spg}
\title{Significant SNP-Peak-Gene}
\usage{
sig_spg(chr, sig_matches)
}
\arguments{
\item{chr}{}

\item{sig_matches}{}
}
\description{
Obtain mediation results for significant SPG match for a chromosome as a list.
 This is needed for sensitivity analysis
}
