% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_readcount_function.R
\name{transpose_readcount}
\alias{transpose_readcount}
\title{Transpose Readcount}
\usage{
transpose_readcount(readcount.path)
}
\arguments{
\item{readcount.path}{ChIP-seq readcount dataframe file path}
}
\description{
Read in and transpose ChIP-seq Readcount dataframe and match project id
}
\keyword{dataprep}
